backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-JWT-Extended
Flask-Mail
python-dotenv
PyMySQL
cryptography
pydantic
email-validator
gunicorn
bcrypt
flask-socketio
twilio
pywebpush
